#491 Implement new card-based grid for displaying templates
