+++ src/components/templates-view/templates-view.tsx
@@ -1,6 +1,6 @@
 import PageHeader from './page-header';
-import TemplatesGrid from '../ag-grid/templates-grid';
 import { RmgPage } from '@railmapgen/rmg-components';
+import TemplatesGrid from './templates-grid';
 
 export default function TemplatesView() {
     return (
